size_t is 32 bits on pi and 64 on mac. int is more consistent.

+++ balloon-communication/udp/test/udp_test_util.h
@@ -8,7 +8,7 @@
 
 struct TestStruct {
     float decimal;
-    size_t string_len;
+    unsigned int string_len;
     char string[MAX_STRING_LEN];
 } TestStruct;
